Check landscape data

Check that landscape inputs have valid values and matching properties.

check(a, b)

# S4 method for RasterLayer,ANY
check(a)

# S4 method for matrix,ANY
check(a)

# S4 method for RasterLayer,RasterLayer
check(a, b)

# S4 method for matrix,matrix
check(a, b)

# S4 method for samc,RasterLayer
check(a, b)

# S4 method for samc,matrix
check(a, b)

Arguments

a
A samc-class, matrix, or RasterLayer-class object

b
A matrix or RasterLayer-class object

Value

See Details section.

Details

This function is used to ensure that landscape inputs (resistance, absorption, fidelity, and occupancy) have valid values and the same properties. This includes checking the CRS (if using RasterLayer inputs), dimensions, and locations of cells with NA data. It can be used to directly compare two matrices or two RasterLayers, or it can be used to check a samc-class object against a matrix or RasterLayer.

The function returns TRUE if the inputs have matching properties. Otherwise, it will stop execution and print the error message generated by the compareRaster() function from the raster package. This error will provide some details about the difference between the two inputs.

Note that the package assumes the different landscape inputs will be the same type, either matrices or RasterLayers. Mixing RasterLayer data and matrix data is not supported.