Get cell numbers

Get cell numbers from raster data

locate(samc, xy)

# S4 method for samc,missing
locate(samc)

# S4 method for samc,ANY
locate(samc, xy)

Arguments

samc

A samc-class object

xy

Any valid input to the y argument of the extract function in the raster package.

Value

A SpatRaster, RasterLayer, matrix, or a vector

Details

This function is used to get cell numbers from raster data. The numbers used for origin and destination values in many samc metrics refer to column/row numbers of the P matrix. For a P matrix derived from raster data, these numbers would normally line up with the cell numbers of the raster, but this is not always true. This is the case when the raster contains NA data; the cells associated with this data are excluded from the P matrix. This causes issues trying to determine the cell numbers that should be used in analyses.

The locate function operates more-or-less like the cellFromXY function in the raster package, but unlike cellFromXY, locate properly accounts for NA cells in identifying cell numbers from coordinate data.

This function can also be used if the samc object was created from matrix inputs for the resistance, absorption, and fidelity parameters. In this case, the values in the xy coordinate parameter can be column-row values with the caveat that (1,1) is the bottom left corner.

The xy parameter can also be excluded. In this case, the function returns a raster where the values of the cells contains the cell number.

Internally, this function relies on the extract function from the raster package, and any valid input for the y argument of that function is valid here.

Examples

library(terra)
#> terra 1.7.3
#> 
#> Attaching package: ‘terra’
#> The following object is masked from ‘package:samc’:
#> 
#>     rasterize
library(samc)

# Load example data
res_data <- samc::example_split_corridor$res
abs_data <- samc::example_split_corridor$abs

# Create samc-class object
samc_obj <- samc(res_data, abs_data,
                 model = list(fun = function(x) 1/mean(x), dir = 8, sym = TRUE))

# We can use locate() to return an object with the cell numbers encoded as data
# in the cells
cell_raster <- locate(samc_obj)
print(cell_raster)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [2,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [3,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [4,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [5,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [6,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [7,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [8,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [9,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [10,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [11,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [12,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [13,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [14,]   NA   NA   NA   NA   NA 1025 1026 1027 1028  1029  1030  1031  1032
#> [15,]   NA   NA   NA   NA   NA 1103 1104 1105 1106  1107  1108  1109  1110
#> [16,]   NA   NA   NA   NA   NA 1177 1178 1179 1180  1181  1182  1183  1184
#> [17,]   NA   NA   NA   NA   NA 1247 1248 1249 1250  1251  1252  1253  1254
#> [18,]   NA   NA   NA   NA   NA 1313 1314 1315 1316  1317  1318  1319  1320
#> [19,]   NA   NA   NA   NA   NA 1379 1380 1381 1382  1383  1384  1385  1386
#> [20,]   NA   NA   NA   NA   NA 1449 1450 1451 1452  1453  1454  1455  1456
#> [21,]   NA   NA   NA   NA   NA 1523 1524 1525 1526  1527  1528  1529  1530
#> [22,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [23,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [24,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [25,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [26,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [27,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [28,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [29,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [30,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [31,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [32,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [33,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#> [34,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [3,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [4,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [5,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [6,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [7,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [8,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [9,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [10,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [11,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [12,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [13,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [14,]  1033  1034  1035  1036  1037  1038  1039  1040  1041  1042  1043  1044
#> [15,]  1111  1112  1113  1114  1115  1116  1117  1118  1119  1120  1121  1122
#> [16,]  1185  1186  1187  1188  1189  1190  1191  1192  1193  1194  1195  1196
#> [17,]  1255  1256  1257  1258  1259  1260  1261  1262  1263  1264  1265  1266
#> [18,]  1321  1322  1323  1324  1325  1326  1327  1328  1329  1330  1331  1332
#> [19,]  1387  1388  1389  1390  1391  1392  1393  1394  1395  1396  1397  1398
#> [20,]  1457  1458  1459  1460  1461  1462  1463  1464  1465  1466  1467  1468
#> [21,]  1531  1532  1533  1534  1535  1536  1537  1538  1539  1540  1541  1542
#> [22,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [23,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [24,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [25,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [26,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [27,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [28,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [29,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [30,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [31,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [32,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [33,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [34,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>       [,26] [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37]
#>  [1,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [3,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [4,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [5,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [6,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [7,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [8,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [9,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [10,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA   481
#> [11,]    NA    NA    NA    NA    NA    NA    NA    NA    NA   611   612   613
#> [12,]    NA    NA    NA    NA    NA    NA    NA   745   746   747   748   749
#> [13,]    NA    NA    NA    NA    NA   883   884   885   886   887   888   889
#> [14,]  1045  1046  1047  1048  1049  1050  1051  1052  1053  1054  1055  1056
#> [15,]  1123  1124  1125  1126  1127  1128  1129  1130  1131  1132  1133  1134
#> [16,]  1197  1198  1199  1200  1201  1202  1203  1204  1205  1206  1207  1208
#> [17,]  1267  1268  1269  1270  1271  1272  1273  1274  1275  1276  1277  1278
#> [18,]  1333  1334  1335  1336  1337  1338  1339  1340  1341  1342  1343  1344
#> [19,]  1399  1400  1401  1402  1403  1404  1405  1406  1407  1408  1409  1410
#> [20,]  1469  1470  1471  1472  1473  1474  1475  1476  1477  1478  1479  1480
#> [21,]  1543  1544  1545  1546  1547  1548  1549  1550  1551  1552  1553  1554
#> [22,]    NA    NA    NA    NA    NA  1601  1602  1603  1604  1605  1606  1607
#> [23,]    NA    NA    NA    NA    NA    NA    NA  1743  1744  1745  1746  1747
#> [24,]    NA    NA    NA    NA    NA    NA    NA    NA    NA  1881  1882  1883
#> [25,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA  2015
#> [26,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [27,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [28,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [29,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [30,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [31,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [32,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [33,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [34,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>       [,38] [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49]
#>  [1,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [3,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [4,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [5,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [6,]    NA    NA    NA    NA    NA    NA    NA     1     2     3     4     5
#>  [7,]    NA    NA    NA    NA    NA   115   116   117   118   119   120   121
#>  [8,]    NA    NA    NA   233   234   235   236   237   238   239   240   241
#>  [9,]    NA   355   356   357   358   359   360   361   362   363   364   365
#> [10,]   482   483   484   485   486   487   488   489   490   491   492   493
#> [11,]   614   615   616   617   618   619   620   621   622   623   624   625
#> [12,]   750   751   752   753   754   755   756   757   758   759   760   761
#> [13,]   890   891   892   893   894   895   896   897   898   899   900   901
#> [14,]  1057  1058  1059  1060  1061  1062  1063    NA    NA    NA    NA    NA
#> [15,]  1135  1136  1137  1138  1139    NA    NA    NA    NA    NA    NA    NA
#> [16,]  1209  1210  1211    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [17,]  1279    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [18,]  1345    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [19,]  1411  1412  1413    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [20,]  1481  1482  1483  1484  1485    NA    NA    NA    NA    NA    NA    NA
#> [21,]  1555  1556  1557  1558  1559  1560  1561    NA    NA    NA    NA    NA
#> [22,]  1608  1609  1610  1611  1612  1613  1614  1615  1616  1617  1618  1619
#> [23,]  1748  1749  1750  1751  1752  1753  1754  1755  1756  1757  1758  1759
#> [24,]  1884  1885  1886  1887  1888  1889  1890  1891  1892  1893  1894  1895
#> [25,]  2016  2017  2018  2019  2020  2021  2022  2023  2024  2025  2026  2027
#> [26,]    NA  2145  2146  2147  2148  2149  2150  2151  2152  2153  2154  2155
#> [27,]    NA    NA    NA  2271  2272  2273  2274  2275  2276  2277  2278  2279
#> [28,]    NA    NA    NA    NA    NA  2393  2394  2395  2396  2397  2398  2399
#> [29,]    NA    NA    NA    NA    NA    NA    NA  2511  2512  2513  2514  2515
#> [30,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [31,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [32,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [33,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [34,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>       [,50] [,51] [,52] [,53] [,54] [,55] [,56] [,57] [,58] [,59] [,60] [,61]
#>  [1,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [3,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [4,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [5,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [6,]     6     7     8     9    10    11    12    13    14    15    16    17
#>  [7,]   122   123   124   125   126   127   128   129   130   131   132   133
#>  [8,]   242   243   244   245   246   247   248   249   250   251   252   253
#>  [9,]   366   367   368   369   370   371   372   373   374   375   376   377
#> [10,]   494   495   496   497   498   499   500   501   502   503   504   505
#> [11,]   626   627   628   629   630   631   632   633   634   635   636   637
#> [12,]   762   763   764   765   766   767   768   769   770   771   772   773
#> [13,]   902   903   904   905   906   907   908   909   910   911   912   913
#> [14,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [15,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [16,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [17,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [18,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [19,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [20,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [21,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [22,]  1620  1621  1622  1623  1624  1625  1626  1627  1628  1629  1630  1631
#> [23,]  1760  1761  1762  1763  1764  1765  1766  1767  1768  1769  1770  1771
#> [24,]  1896  1897  1898  1899  1900  1901  1902  1903  1904  1905  1906  1907
#> [25,]  2028  2029  2030  2031  2032  2033  2034  2035  2036  2037  2038  2039
#> [26,]  2156  2157  2158  2159  2160  2161  2162  2163  2164  2165  2166  2167
#> [27,]  2280  2281  2282  2283  2284  2285  2286  2287  2288  2289  2290  2291
#> [28,]  2400  2401  2402  2403  2404  2405  2406  2407  2408  2409  2410  2411
#> [29,]  2516  2517  2518  2519  2520  2521  2522  2523  2524  2525  2526  2527
#> [30,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [31,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [32,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [33,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [34,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>       [,62] [,63] [,64] [,65] [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73]
#>  [1,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [3,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [4,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [5,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [6,]    18    19    20    21    22    23    24    25    26    27    28    29
#>  [7,]   134   135   136   137   138   139   140   141   142   143   144   145
#>  [8,]   254   255   256   257   258   259   260   261   262   263   264   265
#>  [9,]   378   379   380   381   382   383   384   385   386   387   388   389
#> [10,]   506   507   508   509   510   511   512   513   514   515   516   517
#> [11,]   638   639   640   641   642   643   644   645   646   647   648   649
#> [12,]   774   775   776   777   778   779   780   781   782   783   784   785
#> [13,]   914   915   916   917   918   919   920   921   922   923   924   925
#> [14,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [15,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [16,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [17,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [18,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [19,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [20,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [21,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [22,]  1632  1633  1634  1635  1636  1637  1638  1639  1640  1641  1642  1643
#> [23,]  1772  1773  1774  1775  1776  1777  1778  1779  1780  1781  1782  1783
#> [24,]  1908  1909  1910  1911  1912  1913  1914  1915  1916  1917  1918  1919
#> [25,]  2040  2041  2042  2043  2044  2045  2046  2047  2048  2049  2050  2051
#> [26,]  2168  2169  2170  2171  2172  2173  2174  2175  2176  2177  2178  2179
#> [27,]  2292  2293  2294  2295  2296  2297  2298  2299  2300  2301  2302  2303
#> [28,]  2412  2413  2414  2415  2416  2417  2418  2419  2420  2421  2422  2423
#> [29,]  2528  2529  2530  2531  2532  2533  2534  2535  2536  2537  2538  2539
#> [30,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [31,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [32,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [33,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [34,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>       [,74] [,75] [,76] [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85]
#>  [1,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [3,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [4,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [5,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [6,]    30    31    32    33    34    35    36    37    38    39    40    41
#>  [7,]   146   147   148   149   150   151   152   153   154   155   156   157
#>  [8,]   266   267   268   269   270   271   272   273   274   275   276   277
#>  [9,]   390   391   392   393   394   395   396   397   398   399   400   401
#> [10,]   518   519   520   521   522   523   524   525   526   527   528   529
#> [11,]   650   651   652   653   654   655   656   657   658   659   660   661
#> [12,]   786   787   788   789   790   791   792   793   794   795   796   797
#> [13,]   926   927   928   929   930   931   932   933   934   935   936   937
#> [14,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [15,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [16,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [17,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [18,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [19,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [20,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [21,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [22,]  1644  1645  1646  1647  1648  1649  1650  1651  1652  1653  1654  1655
#> [23,]  1784  1785  1786  1787  1788  1789  1790  1791  1792  1793  1794  1795
#> [24,]  1920  1921  1922  1923  1924  1925  1926  1927  1928  1929  1930  1931
#> [25,]  2052  2053  2054  2055  2056  2057  2058  2059  2060  2061  2062  2063
#> [26,]  2180  2181  2182  2183  2184  2185  2186  2187  2188  2189  2190  2191
#> [27,]  2304  2305  2306  2307  2308  2309  2310  2311  2312  2313  2314  2315
#> [28,]  2424  2425  2426  2427  2428  2429  2430  2431  2432  2433  2434  2435
#> [29,]  2540  2541  2542  2543  2544  2545  2546  2547  2548  2549  2550  2551
#> [30,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [31,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [32,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [33,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [34,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>       [,86] [,87] [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97]
#>  [1,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [3,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [4,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [5,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [6,]    42    43    44    45    46    47    48    49    50    51    52    53
#>  [7,]   158   159   160   161   162   163   164   165   166   167   168   169
#>  [8,]   278   279   280   281   282   283   284   285   286   287   288   289
#>  [9,]   402   403   404   405   406   407   408   409   410   411   412   413
#> [10,]   530   531   532   533   534   535   536   537   538   539   540   541
#> [11,]   662   663   664   665   666   667   668   669   670   671   672   673
#> [12,]   798   799   800   801   802   803   804   805   806   807   808   809
#> [13,]   938   939   940   941   942   943   944   945   946   947   948   949
#> [14,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [15,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [16,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [17,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [18,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [19,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [20,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [21,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [22,]  1656  1657  1658  1659  1660  1661  1662  1663  1664  1665  1666  1667
#> [23,]  1796  1797  1798  1799  1800  1801  1802  1803  1804  1805  1806  1807
#> [24,]  1932  1933  1934  1935  1936  1937  1938  1939  1940  1941  1942  1943
#> [25,]  2064  2065  2066  2067  2068  2069  2070  2071  2072  2073  2074  2075
#> [26,]  2192  2193  2194  2195  2196  2197  2198  2199  2200  2201  2202  2203
#> [27,]  2316  2317  2318  2319  2320  2321  2322  2323  2324  2325  2326  2327
#> [28,]  2436  2437  2438  2439  2440  2441  2442  2443  2444  2445  2446  2447
#> [29,]  2552  2553  2554  2555  2556  2557  2558  2559  2560  2561  2562  2563
#> [30,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [31,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [32,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [33,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [34,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>       [,98] [,99] [,100] [,101] [,102] [,103] [,104] [,105] [,106] [,107]
#>  [1,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [2,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [3,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [4,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [5,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [6,]    54    55     56     57     58     59     60     61     62     63
#>  [7,]   170   171    172    173    174    175    176    177    178    179
#>  [8,]   290   291    292    293    294    295    296    297    298    299
#>  [9,]   414   415    416    417    418    419    420    421    422    423
#> [10,]   542   543    544    545    546    547    548    549    550    551
#> [11,]   674   675    676    677    678    679    680    681    682    683
#> [12,]   810   811    812    813    814    815    816    817    818    819
#> [13,]   950   951    952    953    954    955    956    957    958    959
#> [14,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [15,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [16,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [17,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [18,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [19,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [20,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [21,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [22,]  1668  1669   1670   1671   1672   1673   1674   1675   1676   1677
#> [23,]  1808  1809   1810   1811   1812   1813   1814   1815   1816   1817
#> [24,]  1944  1945   1946   1947   1948   1949   1950   1951   1952   1953
#> [25,]  2076  2077   2078   2079   2080   2081   2082   2083   2084   2085
#> [26,]  2204  2205   2206   2207   2208   2209   2210   2211   2212   2213
#> [27,]  2328  2329   2330   2331   2332   2333   2334   2335   2336   2337
#> [28,]  2448  2449   2450   2451   2452   2453   2454   2455   2456   2457
#> [29,]  2564  2565   2566   2567   2568   2569   2570   2571   2572   2573
#> [30,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [31,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [32,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [33,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [34,]    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#>       [,108] [,109] [,110] [,111] [,112] [,113] [,114] [,115] [,116] [,117]
#>  [1,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [2,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [3,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [4,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [5,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [6,]     64     65     66     67     68     69     70     71     72     73
#>  [7,]    180    181    182    183    184    185    186    187    188    189
#>  [8,]    300    301    302    303    304    305    306    307    308    309
#>  [9,]    424    425    426    427    428    429    430    431    432    433
#> [10,]    552    553    554    555    556    557    558    559    560    561
#> [11,]    684    685    686    687    688    689    690    691    692    693
#> [12,]    820    821    822    823    824    825    826    827    828    829
#> [13,]    960    961    962    963    964    965    966    967    968    969
#> [14,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [15,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [16,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [17,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [18,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [19,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [20,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [21,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [22,]   1678   1679   1680   1681   1682   1683   1684   1685   1686   1687
#> [23,]   1818   1819   1820   1821   1822   1823   1824   1825   1826   1827
#> [24,]   1954   1955   1956   1957   1958   1959   1960   1961   1962   1963
#> [25,]   2086   2087   2088   2089   2090   2091   2092   2093   2094   2095
#> [26,]   2214   2215   2216   2217   2218   2219   2220   2221   2222   2223
#> [27,]   2338   2339   2340   2341   2342   2343   2344   2345   2346   2347
#> [28,]   2458   2459   2460   2461   2462   2463   2464   2465   2466   2467
#> [29,]   2574   2575   2576   2577   2578   2579   2580   2581   2582   2583
#> [30,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [31,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [32,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [33,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [34,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>       [,118] [,119] [,120] [,121] [,122] [,123] [,124] [,125] [,126] [,127]
#>  [1,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [2,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [3,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [4,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [5,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [6,]     74     75     76     77     78     79     80     81     82     83
#>  [7,]    190    191    192    193    194    195    196    197    198    199
#>  [8,]    310    311    312    313    314    315    316    317    318    319
#>  [9,]    434    435    436    437    438    439    440    441    442    443
#> [10,]    562    563    564    565    566    567    568    569    570    571
#> [11,]    694    695    696    697    698    699    700    701    702    703
#> [12,]    830    831    832    833    834    835    836    837    838    839
#> [13,]    970    971    972    973    974    975    976    977    978    979
#> [14,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [15,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [16,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [17,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [18,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [19,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [20,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [21,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [22,]   1688   1689   1690   1691   1692   1693   1694   1695   1696   1697
#> [23,]   1828   1829   1830   1831   1832   1833   1834   1835   1836   1837
#> [24,]   1964   1965   1966   1967   1968   1969   1970   1971   1972   1973
#> [25,]   2096   2097   2098   2099   2100   2101   2102   2103   2104   2105
#> [26,]   2224   2225   2226   2227   2228   2229   2230   2231   2232   2233
#> [27,]   2348   2349   2350   2351   2352   2353   2354   2355   2356   2357
#> [28,]   2468   2469   2470   2471   2472   2473   2474   2475   2476   2477
#> [29,]   2584   2585   2586   2587   2588   2589   2590   2591   2592   2593
#> [30,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [31,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [32,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [33,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [34,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>       [,128] [,129] [,130] [,131] [,132] [,133] [,134] [,135] [,136] [,137]
#>  [1,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [2,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [3,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [4,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [5,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [6,]     84     85     86     87     88     89     90     91     92     93
#>  [7,]    200    201    202    203    204    205    206    207    208    209
#>  [8,]    320    321    322    323    324    325    326    327    328    329
#>  [9,]    444    445    446    447    448    449    450    451    452    453
#> [10,]    572    573    574    575    576    577    578    579    580    581
#> [11,]    704    705    706    707    708    709    710    711    712    713
#> [12,]    840    841    842    843    844    845    846    847    848    849
#> [13,]    980    981    982    983    984    985    986    987    988    989
#> [14,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [15,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [16,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [17,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [18,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [19,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [20,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [21,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [22,]   1698   1699   1700   1701   1702   1703   1704   1705   1706   1707
#> [23,]   1838   1839   1840   1841   1842   1843   1844   1845   1846   1847
#> [24,]   1974   1975   1976   1977   1978   1979   1980   1981   1982   1983
#> [25,]   2106   2107   2108   2109   2110   2111   2112   2113   2114   2115
#> [26,]   2234   2235   2236   2237   2238   2239   2240   2241   2242   2243
#> [27,]   2358   2359   2360   2361   2362   2363   2364   2365   2366   2367
#> [28,]   2478   2479   2480   2481   2482   2483   2484   2485   2486   2487
#> [29,]   2594   2595   2596   2597   2598   2599   2600   2601   2602   2603
#> [30,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [31,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [32,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [33,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [34,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>       [,138] [,139] [,140] [,141] [,142] [,143] [,144] [,145] [,146] [,147]
#>  [1,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [2,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [3,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [4,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [5,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [6,]     94     95     96     97     98     99    100    101    102    103
#>  [7,]    210    211    212    213    214    215    216    217    218    219
#>  [8,]    330    331    332    333    334    335    336    337    338    339
#>  [9,]    454    455    456    457    458    459    460    461    462    463
#> [10,]    582    583    584    585    586    587    588    589    590    591
#> [11,]    714    715    716    717    718    719    720    721    722    723
#> [12,]    850    851    852    853    854    855    856    857    858    859
#> [13,]    990    991    992    993    994    995    996    997    998    999
#> [14,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [15,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [16,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [17,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [18,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [19,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [20,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [21,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [22,]   1708   1709   1710   1711   1712   1713   1714   1715   1716   1717
#> [23,]   1848   1849   1850   1851   1852   1853   1854   1855   1856   1857
#> [24,]   1984   1985   1986   1987   1988   1989   1990   1991   1992   1993
#> [25,]   2116   2117   2118   2119   2120   2121   2122   2123   2124   2125
#> [26,]   2244   2245   2246   2247   2248   2249   2250   2251   2252   2253
#> [27,]   2368   2369   2370   2371   2372   2373   2374   2375   2376   2377
#> [28,]   2488   2489   2490   2491   2492   2493   2494   2495   2496   2497
#> [29,]   2604   2605   2606   2607   2608   2609   2610   2611   2612   2613
#> [30,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [31,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [32,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [33,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [34,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>       [,148] [,149] [,150] [,151] [,152] [,153] [,154] [,155] [,156] [,157]
#>  [1,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [2,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [3,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [4,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [5,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [6,]    104    105    106    107    108    109    110    111    112    113
#>  [7,]    220    221    222    223    224    225    226    227    228    229
#>  [8,]    340    341    342    343    344    345    346    347    348    349
#>  [9,]    464    465    466    467    468    469    470    471    472    473
#> [10,]    592    593    594    595    596    597    598    599    600    601
#> [11,]    724    725    726    727    728    729    730    731    732    733
#> [12,]    860    861    862    863    864    865    866    867    868    869
#> [13,]   1000   1001   1002   1003   1004   1005   1006   1007   1008   1009
#> [14,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [15,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [16,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [17,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [18,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [19,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [20,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [21,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [22,]   1718   1719   1720   1721   1722   1723   1724   1725   1726   1727
#> [23,]   1858   1859   1860   1861   1862   1863   1864   1865   1866   1867
#> [24,]   1994   1995   1996   1997   1998   1999   2000   2001   2002   2003
#> [25,]   2126   2127   2128   2129   2130   2131   2132   2133   2134   2135
#> [26,]   2254   2255   2256   2257   2258   2259   2260   2261   2262   2263
#> [27,]   2378   2379   2380   2381   2382   2383   2384   2385   2386   2387
#> [28,]   2498   2499   2500   2501   2502   2503   2504   2505   2506   2507
#> [29,]   2614   2615   2616   2617   2618   2619   2620   2621   2622   2623
#> [30,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [31,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [32,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [33,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [34,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>       [,158] [,159] [,160] [,161] [,162] [,163] [,164] [,165] [,166] [,167]
#>  [1,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [2,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [3,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [4,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [5,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [6,]    114     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [7,]    230    231    232     NA     NA     NA     NA     NA     NA     NA
#>  [8,]    350    351    352    353    354     NA     NA     NA     NA     NA
#>  [9,]    474    475    476    477    478    479    480     NA     NA     NA
#> [10,]    602    603    604    605    606    607    608    609    610     NA
#> [11,]    734    735    736    737    738    739    740    741    742    743
#> [12,]    870    871    872    873    874    875    876    877    878    879
#> [13,]   1010   1011   1012   1013   1014   1015   1016   1017   1018   1019
#> [14,]     NA   1064   1065   1066   1067   1068   1069   1070   1071   1072
#> [15,]     NA     NA     NA   1140   1141   1142   1143   1144   1145   1146
#> [16,]     NA     NA     NA     NA     NA   1212   1213   1214   1215   1216
#> [17,]     NA     NA     NA     NA     NA     NA     NA   1280   1281   1282
#> [18,]     NA     NA     NA     NA     NA     NA     NA   1346   1347   1348
#> [19,]     NA     NA     NA     NA     NA   1414   1415   1416   1417   1418
#> [20,]     NA     NA     NA   1486   1487   1488   1489   1490   1491   1492
#> [21,]     NA   1562   1563   1564   1565   1566   1567   1568   1569   1570
#> [22,]   1728   1729   1730   1731   1732   1733   1734   1735   1736   1737
#> [23,]   1868   1869   1870   1871   1872   1873   1874   1875   1876   1877
#> [24,]   2004   2005   2006   2007   2008   2009   2010   2011   2012   2013
#> [25,]   2136   2137   2138   2139   2140   2141   2142   2143   2144     NA
#> [26,]   2264   2265   2266   2267   2268   2269   2270     NA     NA     NA
#> [27,]   2388   2389   2390   2391   2392     NA     NA     NA     NA     NA
#> [28,]   2508   2509   2510     NA     NA     NA     NA     NA     NA     NA
#> [29,]   2624     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [30,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [31,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [32,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [33,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [34,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>       [,168] [,169] [,170] [,171] [,172] [,173] [,174] [,175] [,176] [,177]
#>  [1,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [2,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [3,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [4,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [5,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [6,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [7,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [8,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [9,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [10,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [11,]    744     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [12,]    880    881    882     NA     NA     NA     NA     NA     NA     NA
#> [13,]   1020   1021   1022   1023   1024     NA     NA     NA     NA     NA
#> [14,]   1073   1074   1075   1076   1077   1078   1079   1080   1081   1082
#> [15,]   1147   1148   1149   1150   1151   1152   1153   1154   1155   1156
#> [16,]   1217   1218   1219   1220   1221   1222   1223   1224   1225   1226
#> [17,]   1283   1284   1285   1286   1287   1288   1289   1290   1291   1292
#> [18,]   1349   1350   1351   1352   1353   1354   1355   1356   1357   1358
#> [19,]   1419   1420   1421   1422   1423   1424   1425   1426   1427   1428
#> [20,]   1493   1494   1495   1496   1497   1498   1499   1500   1501   1502
#> [21,]   1571   1572   1573   1574   1575   1576   1577   1578   1579   1580
#> [22,]   1738   1739   1740   1741   1742     NA     NA     NA     NA     NA
#> [23,]   1878   1879   1880     NA     NA     NA     NA     NA     NA     NA
#> [24,]   2014     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [25,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [26,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [27,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [28,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [29,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [30,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [31,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [32,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [33,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [34,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>       [,178] [,179] [,180] [,181] [,182] [,183] [,184] [,185] [,186] [,187]
#>  [1,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [2,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [3,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [4,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [5,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [6,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [7,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [8,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [9,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [10,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [11,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [12,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [13,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [14,]   1083   1084   1085   1086   1087   1088   1089   1090   1091   1092
#> [15,]   1157   1158   1159   1160   1161   1162   1163   1164   1165   1166
#> [16,]   1227   1228   1229   1230   1231   1232   1233   1234   1235   1236
#> [17,]   1293   1294   1295   1296   1297   1298   1299   1300   1301   1302
#> [18,]   1359   1360   1361   1362   1363   1364   1365   1366   1367   1368
#> [19,]   1429   1430   1431   1432   1433   1434   1435   1436   1437   1438
#> [20,]   1503   1504   1505   1506   1507   1508   1509   1510   1511   1512
#> [21,]   1581   1582   1583   1584   1585   1586   1587   1588   1589   1590
#> [22,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [23,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [24,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [25,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [26,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [27,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [28,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [29,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [30,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [31,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [32,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [33,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [34,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>       [,188] [,189] [,190] [,191] [,192] [,193] [,194] [,195] [,196] [,197]
#>  [1,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [2,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [3,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [4,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [5,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [6,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [7,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [8,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  [9,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [10,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [11,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [12,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [13,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [14,]   1093   1094   1095   1096   1097   1098   1099   1100   1101   1102
#> [15,]   1167   1168   1169   1170   1171   1172   1173   1174   1175   1176
#> [16,]   1237   1238   1239   1240   1241   1242   1243   1244   1245   1246
#> [17,]   1303   1304   1305   1306   1307   1308   1309   1310   1311   1312
#> [18,]   1369   1370   1371   1372   1373   1374   1375   1376   1377   1378
#> [19,]   1439   1440   1441   1442   1443   1444   1445   1446   1447   1448
#> [20,]   1513   1514   1515   1516   1517   1518   1519   1520   1521   1522
#> [21,]   1591   1592   1593   1594   1595   1596   1597   1598   1599   1600
#> [22,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [23,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [24,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [25,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [26,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [27,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [28,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [29,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [30,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [31,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [32,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [33,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#> [34,]     NA     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>       [,198] [,199] [,200] [,201] [,202]
#>  [1,]     NA     NA     NA     NA     NA
#>  [2,]     NA     NA     NA     NA     NA
#>  [3,]     NA     NA     NA     NA     NA
#>  [4,]     NA     NA     NA     NA     NA
#>  [5,]     NA     NA     NA     NA     NA
#>  [6,]     NA     NA     NA     NA     NA
#>  [7,]     NA     NA     NA     NA     NA
#>  [8,]     NA     NA     NA     NA     NA
#>  [9,]     NA     NA     NA     NA     NA
#> [10,]     NA     NA     NA     NA     NA
#> [11,]     NA     NA     NA     NA     NA
#> [12,]     NA     NA     NA     NA     NA
#> [13,]     NA     NA     NA     NA     NA
#> [14,]     NA     NA     NA     NA     NA
#> [15,]     NA     NA     NA     NA     NA
#> [16,]     NA     NA     NA     NA     NA
#> [17,]     NA     NA     NA     NA     NA
#> [18,]     NA     NA     NA     NA     NA
#> [19,]     NA     NA     NA     NA     NA
#> [20,]     NA     NA     NA     NA     NA
#> [21,]     NA     NA     NA     NA     NA
#> [22,]     NA     NA     NA     NA     NA
#> [23,]     NA     NA     NA     NA     NA
#> [24,]     NA     NA     NA     NA     NA
#> [25,]     NA     NA     NA     NA     NA
#> [26,]     NA     NA     NA     NA     NA
#> [27,]     NA     NA     NA     NA     NA
#> [28,]     NA     NA     NA     NA     NA
#> [29,]     NA     NA     NA     NA     NA
#> [30,]     NA     NA     NA     NA     NA
#> [31,]     NA     NA     NA     NA     NA
#> [32,]     NA     NA     NA     NA     NA
#> [33,]     NA     NA     NA     NA     NA
#> [34,]     NA     NA     NA     NA     NA

# We can use a variety of spatial inputs to get cell numbers using locate()
# The simplest is a two-column data.frame
coords <- data.frame(x = c(50, 79, 22),
                     y = c(25, 11, 19))
print(coords)
#>    x  y
#> 1 50 25
#> 2 79 11
#> 3 22 19
locate(samc_obj, coords)
#> [1]  494 1925 1193

# You will get an error if you input a coordinate that does not correspond
# to a non-NA cell
coords <- data.frame(x = c(1),
                     y = c(1))
print(coords)
#>   x y
#> 1 1 1
try(locate(samc_obj, coords))
#> Error : One or more coordinates do not correspond to non-NA cells.