Calculate mortality metrics

Calculates the probability of experiencing mortality at specific locations.

mortality(samc, occ, origin, dest, time)

# S4 method for samc,missing,missing,missing,numeric
mortality(samc, time)

# S4 method for samc,missing,numeric,missing,numeric
mortality(samc, origin,
  time)

# S4 method for samc,missing,missing,numeric,numeric
mortality(samc, dest,
  time)

# S4 method for samc,missing,numeric,numeric,numeric
mortality(samc, origin,
  dest, time)

# S4 method for samc,RasterLayer,missing,missing,numeric
mortality(samc, occ,
  time)

# S4 method for samc,matrix,missing,missing,numeric
mortality(samc, occ, time)

# S4 method for samc,missing,missing,missing,missing
mortality(samc)

# S4 method for samc,missing,numeric,missing,missing
mortality(samc, origin)

# S4 method for samc,missing,missing,numeric,missing
mortality(samc, dest)

# S4 method for samc,missing,numeric,numeric,missing
mortality(samc, origin,
  dest)

# S4 method for samc,RasterLayer,missing,missing,missing
mortality(samc, occ)

# S4 method for samc,matrix,missing,missing,missing
mortality(samc, occ)

Arguments

samc
A samc-class object. This should be output from the samc function.

occ
A RasterLayer-class or matrix. The input type must match the input type used to create the samc-class object, and must have the same properties as the rest of the landscape data. See the check function for more details.

origin
A positive integer representing a cell in the landscape, excluding NA cells. Corresponds to row i of matrix P in the samc-class object.

dest
A positive integer representing a cell in the landscape, excluding NA cells. Corresponds to column j of matrix P in the samc-class object.

time
A positive integer representing time steps

Value

A matrix, vector, or numeric

Details

\(\tilde{B}_t = (\sum_{n=0}^{t-1} Q^n) \tilde{R}\)

mortality(samc, time)

The result is a matrix where element (i,j) is the probability of experiencing mortality at location j within t or fewer steps if starting at location i.

The returned matrix will always be dense and cannot be optimized. Must enable override to use.

mortality(samc, origin, time)

The result is a vector where each element corresponds to a cell in the landscape, and can be mapped back to the landscape using the map function. Element j is the probability of experiencing mortality at location j within t or fewer steps if starting at a given origin.

mortality(samc, dest, time)

The result is a vector where each element corresponds to a cell in the landscape, and can be mapped back to the landscape using the map function. Element i is the probability of experiencing mortality at a given destination within t or fewer steps if starting at location i.

mortality(samc, origin, dest, time)

The result is a numeric value that is the probability of experiencing mortality at a given destination within t or fewer steps if starting at a given origin.

\(\psi^T \tilde{B}_t\)

mortality(samc, occ, time)

The result is a vector where each element corresponds to a cell in the landscape, and can be mapped back to the landscape using the map function. Element j is the unconditional probability of experiencing mortality at location j within t or fewer time steps.

\(B = F \tilde{R}\)

mortality(samc)

The result is a matrix where element (i,j) is the probability of experiencing mortality at location j if starting at location i.

The returned matrix will always be dense and cannot be optimized. Must enable override to use.

mortality(samc, origin)

The result is a vector where each element corresponds to a cell in the landscape, and can be mapped back to the landscape using the map function. Element j is the probability of experiencing mortality at location j if starting at a givin origin.

mortality(samc, dest)

The result is a vector where each element corresponds to a cell in the landscape, and can be mapped back to the landscape using the map function. Element i is the probability of experiencing mortality at a given destination if starting at location i.

mortality(samc, origin, dest)

The result is a numeric value that is the probability of experiencing mortality at a given destination if starting at a given origin

\(\psi^T B\)

mortality(samc, occ)

The result is a vector where each element corresponds to a cell in the landscape, and can be mapped back to the landscape using the map function. Element j is the unconditional probability of experiencing mortality at location j, regardless of the initial state.

Performance

Any relevant performance information about this function can be found in the performance vignette: vignette("performance", package = "samc")