Calculate visitation metrics

Calculates the number of times that individuals from each location visit each location in the landscape before death.

visitation(samc, origin, dest)

# S4 method for samc,missing,missing
visitation(samc)

# S4 method for samc,numeric,missing
visitation(samc, origin)

# S4 method for samc,missing,numeric
visitation(samc, dest)

# S4 method for samc,numeric,numeric
visitation(samc, origin, dest)

Arguments

samc
A samc-class object. This should be output from the samc function.

origin
A positive integer representing a cell in the landscape, excluding NA cells. Corresponds to row i of matrix P in the samc-class object.

dest
A positive integer representing a cell in the landscape, excluding NA cells. Corresponds to column j of matrix P in the samc-class object.

Value

A matrix, a vector, or a numeric

Details

\(F = (I-Q)^{-1}\)

visitation(samc)

The result is a matrix where element (i,j) is the expected number of times an individual that starts in i uses j before it dies.

The returned matrix will always be dense and cannot be optimized. Must enable override to use.

visitation(samc, origin)

The result is a vector where each element corresponds to a cell in the landscape, and can be mapped back to the landscape using the map function. Element j is the number of times that an individual starting at the origin visits location j before it dies.

visitation(samc, dest)

The result is a vector where each element corresponds to a cell in the landscape, and can be mapped back to the landscape using the map function. Element i is the number of times that an individual starting at location i visits the destination before it dies.

visitation(samc, origin, dest)

The result is a numeric value that is the expected number of times an individual starting at the origin visits the destination before it dies.

Performance

Any relevant performance information about this function can be found in the performance vignette: vignette("performance", package = "samc")

Examples

# "Load" the data. In this case we are using data built into the package.
# In practice, users will likely load raster data using the raster() function
# from the raster package.
res_data <- samc::ex_res_data
abs_data <- samc::ex_abs_data
occ_data <- samc::ex_occ_data

# Make sure our data meets the basic input requirements of the package using
# the check() function.
check(res_data, abs_data)

#> [1] TRUE

check(res_data, occ_data)

#> [1] TRUE

# Create a `samc-class` object with the resistance and absorption data using
# the samc() function. We use the recipricol of the arithmetic mean for
# calculating the transition matrix. Note, the input data here are matrices,
# not RasterLayers. If using RasterLayers, the latlon parameter must be set.
samc_obj <- samc(res_data, abs_data, tr_fun = function(x) 1/mean(x))

# Convert the occupancy data to probability of occurrence
occ_prob_data <- occ_data / sum(occ_data, na.rm = TRUE)

# Calculate short- and long-term metrics using the analytical functions
short_mort <- mortality(samc_obj, occ_prob_data, time = 50)
short_dist <- distribution(samc_obj, origin = 3, time = 50)
long_disp <- dispersal(samc_obj, occ_prob_data)

#> Performing setup. This can take several minutes... Complete.
#> Calculating matrix inverse diagonal...Calculating matrix inverse diagonal... Complete                                           
#> Performing final calculations. This may take a few minutes... Complete.

visit <- visitation(samc_obj, dest = 4)
surv <- survival(samc_obj)

# Use the map() function to turn vector results into RasterLayer objects.
short_mort_map <- map(samc_obj, short_mort)
short_dist_map <- map(samc_obj, short_dist)
long_disp_map <- map(samc_obj, long_disp)
visit_map <- map(samc_obj, visit)
surv_map <- map(samc_obj, surv)