Calculate visitation metrics

Calculates the number of times that individuals from each location visit each location in the landscape before death.

visitation(samc, origin, dest)

# S4 method for samc,missing,missing
visitation(samc)

# S4 method for samc,numeric,missing
visitation(samc, origin)

# S4 method for samc,missing,numeric
visitation(samc, dest)

# S4 method for samc,numeric,numeric
visitation(samc, origin, dest)

Arguments

samc
A samc-class object. This should be output from the samc function.

origin
A positive integer representing a cell in the landscape, excluding NA cells. Corresponds to row i of matrix P in the samc-class object.

dest
A positive integer representing a cell in the landscape, excluding NA cells. Corresponds to column j of matrix P in the samc-class object.

Value

A matrix, a vector, or a numeric

Details

\(F = (I-Q)^{-1}\)

visitation(samc)

The result is a matrix where element (i,j) is the expected number of times an individual that starts in i uses j before it dies.

The returned matrix will always be dense and cannot be optimized. Must enable override to use.

visitation(samc, origin)

The result is a vector where each element corresponds to a cell in the landscape, and can be mapped back to the landscape using the map function. Element j is the number of times that an individual starting at the origin visits location j before it dies.

visitation(samc, dest)

The result is a vector where each element corresponds to a cell in the landscape, and can be mapped back to the landscape using the map function. Element i is the number of times that an individual starting at location i visits the destination before it dies.

visitation(samc, origin, dest)

The result is a numeric value that is the expected number of times an individual starting at the origin visits the destination before it dies.

Performance

Any relevant performance information about this function can be found in the performance vignette: vignette("performance", package = "samc")